refactor(canister-core): consolidate all impl State blocks under state/
Every `impl State` block now lives in the `state/` module, split across
focused submodules (assets, rules, env, hashing, serving, sync, protection,
upgrade), so State's whole behavioral surface is discoverable in one place.
Previously the blocks were scattered across state.rs, serving.rs, sync.rs, and
protection/. The domain data types stay in their own lower-layer modules
(protection.rs, and sync's types in state/sync.rs).

Because all impl blocks are now descendants of `state`, the owned pieces
`store`/`certifier`/`sync_session`/`token_index` drop from pub(crate) to
private — nothing outside state/ can reach into them. Only `chunks` stays
pub(crate) (tests/benches stage chunks directly). The `sync` data types cross
the boundary as `crate::state::sync::…`.

Also drops the "gate" terminology throughout docs/comments in favor of the
single term "access protection" (leaving the unrelated HTTP "gateway"
references intact).
